Separate session_page into ssh_session_page and sftp_session_page
