+++ frontend/lib/modules/session/ui/session_title_bar.dart
@@ -11,7 +11,7 @@ import '../../connections/provider/connection.provider.dart';
 import '../../connections/ui/connection_icon.dart';
 import '../../settings/provider/terminal_theme.provider.dart';
 import '../provider/session.provider.dart';
-import '../view/session_page.dart';
+import '../view/ssh_session_page.dart';
 
 class SessionTitleBar extends HookConsumerWidget {
   final String sessionId;
